requirements.txt
streamlit==1.45.1
tensorflow==2.16.2
numpy==1.26.4
pandas==2.2.3
matplotlib==3.10.1
seaborn==0.13.2
pillow==11.2.1
scikit-learn==1.6.1
